rust: Move history tree FFI logic into a module
